fix: split work CLI registration into root_work.go for backend-size
