+++ pkg/cli/root_work.go
@@ -1,0 +1,87 @@
+package cli
+
+import (
+	workcli "github.com/portpowered/infinite-you/pkg/cli/work"
+	"github.com/spf13/cobra"
+)
+
+func newWorkCommand(globals *cliGlobalOptions, diagnostics *cliDiagnosticsOptions) *cobra.Command {
+	workCmd := &cobra.Command{
+		Use:   "work",
+		Short: "Inspect work from a running factory",
+	}
+	workCmd.AddCommand(newWorkListCommand(globals, diagnostics))
+	workCmd.AddCommand(newWorkShowCommand(globals, diagnostics))
+	return workCmd
+}
+
+func newWorkListCommand(globals *cliGlobalOptions, diagnostics *cliDiagnosticsOptions) *cobra.Command {
+	cfg := workcli.ListConfig{Server: globals.server}
+
+	cmd := &cobra.Command{
+		Use:   "list",
+		Short: "List work from a running factory",
+		Long: "List work from a running you-agent-factory service.\n\n" +
+			"By default the command targets the default compatibility session. " +
+			"Use --session to route the request to one specific live factory session instead. " +
+			"Run " + cliBinaryName + " session list to discover live session ids.\n\n" +
+			"Optional filters are applied on the server before pagination. " +
+			"--name matches a case-insensitive substring of the work name. " +
+			"--work-type-name matches workTypeName exactly. " +
+			"--trace-id matches traceId or currentChainingTraceId exactly. " +
+			"Combine filters with --max-results and --next-token to page through the filtered result set.",
+		PreRunE: rejectDeprecatedPortFlag,
+		RunE: func(cmd *cobra.Command, args []string) error {
+			cfg.Server = globals.server
+			cfg.JSON = globals.json
+			cfg.Output = cmd.OutOrStdout()
+			cfg.Diagnostics = diagnostics.writer(cmd)
+			cfg.Verbose = diagnostics.verboseEnabled()
+			cfg.Debug = diagnostics.debug
+			return listWork(cfg)
+		},
+	}
+
+	registerDeprecatedPortFlag(cmd)
+	cmd.Flags().StringVar(&cfg.StateName, "state-name", "", "filter by current state name")
+	cmd.Flags().StringVar(&cfg.StateType, "state-type", "", "filter by current state type (INITIAL, PROCESSING, TERMINAL, FAILED)")
+	cmd.Flags().StringVar(&cfg.Name, "name", "", "filter by case-insensitive substring of work name (applied before pagination)")
+	cmd.Flags().StringVar(&cfg.WorkTypeName, "work-type-name", "", "filter by exact workTypeName (applied before pagination)")
+	cmd.Flags().StringVar(&cfg.TraceID, "trace-id", "", "filter by exact traceId or currentChainingTraceId (applied before pagination)")
+	cmd.Flags().StringVar(&cfg.SortBy, "sort-by", "", "sort returned work by field (state.type)")
+	cmd.Flags().IntVar(&cfg.MaxResults, "max-results", 0, "maximum work items to return per page after server-side filters")
+	cmd.Flags().StringVar(&cfg.NextToken, "next-token", "", "pagination cursor returned by a previous work list response")
+	cmd.Flags().StringVar(&cfg.SessionID, "session", "", "target one live factory session; omit to use the default compatibility session")
+	return cmd
+}
+
+func newWorkShowCommand(globals *cliGlobalOptions, diagnostics *cliDiagnosticsOptions) *cobra.Command {
+	cfg := workcli.ShowConfig{Server: globals.server}
+
+	cmd := &cobra.Command{
+		Use:   "show <work-id>",
+		Short: "Show one work item from a running factory",
+		Long: "Show one work item from a running you-agent-factory service.\n\n" +
+			"By default the command targets the default compatibility session. " +
+			"Use --session to route the request to one specific live factory session instead. " +
+			"Run " + cliBinaryName + " session list to discover live session ids.\n\n" +
+			"After submit, use " + cliBinaryName + " work list --name <name> to find work ids, " +
+			"then " + cliBinaryName + " work show <work-id> to verify one item without JSON pagination.",
+		Args:    cobra.ExactArgs(1),
+		PreRunE: rejectDeprecatedPortFlag,
+		RunE: func(cmd *cobra.Command, args []string) error {
+			cfg.Server = globals.server
+			cfg.WorkID = args[0]
+			cfg.JSON = globals.json
+			cfg.Output = cmd.OutOrStdout()
+			cfg.Diagnostics = diagnostics.writer(cmd)
+			cfg.Verbose = diagnostics.verboseEnabled()
+			cfg.Debug = diagnostics.debug
+			return showWork(cfg)
+		},
+	}
+
+	registerDeprecatedPortFlag(cmd)
+	cmd.Flags().StringVar(&cfg.SessionID, "session", "", "target one live factory session; omit to use the default compatibility session")
+	return cmd
+}
